refactor(frontend): enhance article rendering with new font size constants and PostRenderer component
- Introduced ARTICLE_FONT_SIZE_CLASSNAMES and ARTICLE_FONT_SIZE_CLASSNAMES_LARGE for improved font size management.
- Updated ArticleSummary to utilize new font size constants for better styling.
- Moved PostRenderer component to components folder.
